fix: require and import cycles
