+++ package/src/components/AttachmentPicker/components/AttachmentMediaPicker/AttachmentMediaPicker.tsx
@@ -2,6 +2,7 @@ import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { Linking, StyleSheet } from 'react-native';
 
 import { renderAttachmentPickerItem } from './AttachmentPickerItem';
+import { IOS_LIMITED_DEEPLINK, type PhotoContentItemType } from './shared';
 
 import { useAttachmentPickerContext, useTheme, useTranslationContext } from '../../../../contexts';
 
@@ -9,21 +10,11 @@ import { useStableCallback } from '../../../../hooks';
 import { Picture } from '../../../../icons';
 
 import { NativeHandlers } from '../../../../native';
-import type { File } from '../../../../types/types';
 import { BottomSheetFlatList } from '../../../BottomSheetCompatibility/BottomSheetFlatList';
 import {
-  AttachmentPickerContentProps,
   AttachmentPickerGenericContent,
-} from '../AttachmentPickerContent';
-
-export const IOS_LIMITED_DEEPLINK = '@getstream/ios-limited-button' as const;
-
-export type IosLimitedItemType = { uri: typeof IOS_LIMITED_DEEPLINK };
-
-export type PhotoContentItemType = File | IosLimitedItemType;
-
-export const isIosLimited = (item: PhotoContentItemType): item is IosLimitedItemType =>
-  'uri' in item && item.uri === '@getstream/ios-limited-button';
+  type AttachmentPickerContentProps,
+} from '../AttachmentPickerGenericContent';
 
 const keyExtractor = (item: PhotoContentItemType) => item.uri;
